shared-backend/requirements.txt
# Core FastAPI dependencies
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# AI and ML
anthropic==0.7.7
openai==1.3.7
tiktoken==0.5.2
numpy==1.24.4
scikit-learn==1.3.2

# Vector database (for embeddings)
pgvector==0.2.4

# Authentication and security
bcrypt==4.1.2
python-jose==3.3.0
cryptography==41.0.8

# HTTP client
httpx==0.25.2
requests==2.31.0

# Data validation and serialization
pydantic==2.5.0
pydantic-settings==2.1.0

# Background tasks
celery==5.3.4
redis==5.0.1

# File handling
python-magic==0.4.27
Pillow==10.1.0
PyPDF2==3.0.1

# Utilities
python-dotenv==1.0.0
loguru==0.7.2
pydantic-extra-types==2.3.0

# WebSocket support
websockets==12.0

# Email
emails==0.6.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2

# Development
black==23.11.0
isort==5.12.0
flake8==6.1.0
mypy==1.7.1
